fix: moving Cryptography helper to a service
