--- Source/Cogworks.Essentials/Helpers/Interfaces/ICryptographyHelper.cs
@@ -1,7 +1,0 @@
-﻿namespace Cogworks.Essentials.Helpers.Interfaces
-{
-    public interface ICryptographyHelper
-    {
-        string GenerateRandomToken();
-    }
-}
