+++ Source/Cogworks.Essentials/Services/Interfaces/ICryptographyService.cs
@@ -1,0 +1,7 @@
+﻿namespace Cogworks.Essentials.Services.Interfaces
+{
+    public interface ICryptographyService
+    {
+        string GenerateRandomToken();
+    }
+}
